+++ src/main.cpp
@@ -8,6 +8,9 @@
 #include "data/DataStructures.h"
 #include "data/DataProcessing.h"
 
+// Protocol parsers
+#include "parsers/Parsers.h" 
+
 void rpcWrapper() {
   JsonDocument jsonResponse;
   JsonDocument doc;
@@ -415,25 +418,6 @@ void webSocketEvent(AsyncWebSocket *server, AsyncWebSocketClient *client, AwsEve
   }
 }
 
-void mqtt_callback(char *topic, byte *payload, unsigned int length) {
-  JsonDocument json;
-  deserializeJson(json, payload, length);
-  setJsonPathPower(json);
-}
-
-void mqtt_reconnect() {
-  DEBUG_SERIAL.print("Attempting MQTT connection...");
-  if (mqtt_client.connect(shelly_name, String(mqtt_user).c_str(), String(mqtt_passwd).c_str())) {
-    DEBUG_SERIAL.println("connected");
-    mqtt_client.subscribe(mqtt_topic);
-  } else {
-    DEBUG_SERIAL.print("failed, rc=");
-    DEBUG_SERIAL.print(mqtt_client.state());
-    DEBUG_SERIAL.println(" try again in 5 seconds");
-    delay(5000);
-  }
-}
-
 void parseUdpRPC() {
   uint8_t buffer[1024];
   int packetSize = UdpRPC.parsePacket();
@@ -526,299 +510,6 @@ void parseHttpRPC(String requestBody, AsyncWebServerRequest *request) {
       }
     }
   }
-}
-
-void parseSMA() {
-  uint8_t buffer[1024];
-  int packetSize = Udp.parsePacket();
-  if (packetSize) {
-    int rSize = Udp.read(buffer, 1024);
-    if (buffer[0] != 'S' || buffer[1] != 'M' || buffer[2] != 'A') {
-      DEBUG_SERIAL.println("Not an SMA packet?");
-      return;
-    }
-    uint16_t grouplen;
-    uint16_t grouptag;
-    uint8_t *offset = buffer + 4;
-    do {
-      grouplen = (offset[0] << 8) + offset[1];
-      grouptag = (offset[2] << 8) + offset[3];
-      offset += 4;
-      if (grouplen == 0xffff) return;
-      if (grouptag == 0x02A0 && grouplen == 4) {
-        offset += 4;
-      } else if (grouptag == 0x0010) {
-        uint8_t *endOfGroup = offset + grouplen;
-        // uint16_t protocolID = (offset[0] << 8) + offset[1];
-        offset += 2;
-        // uint16_t susyID = (offset[0] << 8) + offset[1];
-        offset += 2;
-        uint32_t serial = (offset[0] << 24) + (offset[1] << 16) + (offset[2] << 8) + offset[3];
-        DEBUG_SERIAL.print("Received SMA multicast from ");
-        DEBUG_SERIAL.println(serial);
-        if ((strcmp(sma_id, "") != 0) && (String(sma_id).toInt() != serial)) {
-          DEBUG_SERIAL.println("SMA serial not matching - ignoring packet");
-          break;
-        }
-        offset += 4;
-        // uint32_t timestamp = (offset[0] << 24) + (offset[1] << 16) + (offset[2] << 8) + offset[3];
-        offset += 4;
-        while (offset < endOfGroup) {
-          uint8_t channel = offset[0];
-          uint8_t index = offset[1];
-          uint8_t type = offset[2];
-          // uint8_t tarif = offset[3];
-          offset += 4;
-          if (type == 8) {
-            uint64_t data = ((uint64_t)offset[0] << 56) + ((uint64_t)offset[1] << 48) + ((uint64_t)offset[2] << 40) + ((uint64_t)offset[3] << 32) + ((uint64_t)offset[4] << 24) + ((uint64_t)offset[5] << 16) + ((uint64_t)offset[6] << 8) + offset[7];
-            offset += 8;
-            switch (index) {
-              case 21:
-                PhaseEnergy[0].consumption = data / 3600000;
-                break;
-              case 22:
-                PhaseEnergy[0].gridfeedin = data / 3600000;
-                break;
-              case 41:
-                PhaseEnergy[1].consumption = data / 3600000;
-                break;
-              case 42:
-                PhaseEnergy[1].gridfeedin = data / 3600000;
-                break;
-              case 61:
-                PhaseEnergy[2].consumption = data / 3600000;
-                break;
-              case 62:
-                PhaseEnergy[2].gridfeedin = data / 3600000;
-                break;
-            }
-          } else if (type == 4) {
-            uint32_t data = (offset[0] << 24) + (offset[1] << 16) + (offset[2] << 8) + offset[3];
-            offset += 4;
-            switch (index) {
-              case 1:
-                // 1.4.0 Total grid power in dW - unused
-                break;
-              case 2:
-                // 2.4.0 Total feed-in power in dW - unused
-                break;
-              case 21:
-                PhasePower[0].power = round2(data * 0.1);
-                PhasePower[0].frequency = defaultFrequency;
-                break;
-              case 22:
-                PhasePower[0].power -= round2(data * 0.1);
-                break;
-              case 29:
-                PhasePower[0].apparentPower = round2(data * 0.1);
-                break;
-              case 30:
-                PhasePower[0].apparentPower -= round2(data * 0.1);
-                break;
-              case 31:
-                PhasePower[0].current = round2(data * 0.001);
-                break;
-              case 32:
-                PhasePower[0].voltage = round2(data * 0.001);
-                break;
-              case 33:
-                PhasePower[0].powerFactor = round2(data * 0.001);
-                break;
-              case 41:
-                PhasePower[1].power = round2(data * 0.1);
-                PhasePower[1].frequency = defaultFrequency;
-                break;
-              case 42:
-                PhasePower[1].power -= round2(data * 0.1);
-                break;
-              case 49:
-                PhasePower[1].apparentPower = round2(data * 0.1);
-                break;
-              case 50:
-                PhasePower[1].apparentPower -= round2(data * 0.1);
-                break;
-              case 51:
-                PhasePower[1].current = round2(data * 0.001);
-                break;
-              case 52:
-                PhasePower[1].voltage = round2(data * 0.001);
-                break;
-              case 53:
-                PhasePower[1].powerFactor = round2(data * 0.001);
-                break;
-              case 61:
-                PhasePower[2].power = round2(data * 0.1);
-                PhasePower[2].frequency = defaultFrequency;
-                break;
-              case 62:
-                PhasePower[2].power -= round2(data * 0.1);
-                break;
-              case 69:
-                PhasePower[2].apparentPower = round2(data * 0.1);
-                break;
-              case 70:
-                PhasePower[2].apparentPower -= round2(data * 0.1);
-                break;
-              case 71:
-                PhasePower[2].current = round2(data * 0.001);
-                break;
-              case 72:
-                PhasePower[2].voltage = round2(data * 0.001);
-                break;
-              case 73:
-                PhasePower[2].powerFactor = round2(data * 0.001);
-                break;
-              default:
-                break;
-            }
-          } else if (channel == 144) {
-            // optional handling of version number
-            offset += 4;
-          } else {
-            offset += type;
-            DEBUG_SERIAL.println("Unknown measurement");
-          }
-        }
-      } else if (grouptag == 0) {
-        // end marker
-        offset += grouplen;
-      } else {
-        DEBUG_SERIAL.print("unhandled group ");
-        DEBUG_SERIAL.print(grouptag);
-        DEBUG_SERIAL.print(" with len=");
-        DEBUG_SERIAL.println(grouplen);
-        offset += grouplen;
-      }
-    } while (grouplen > 0 && offset + 4 < buffer + rSize);
-  }
-}
-
-void parseSHRDZM() {
-  JsonDocument json;
-  uint8_t buffer[1024];
-  int packetSize = Udp.parsePacket();
-  if (packetSize) {
-    int rSize = Udp.read(buffer, 1024);
-    buffer[rSize] = 0;
-    deserializeJson(json, buffer);
-    if (json["data"]["16.7.0"].is<JsonVariant>()) {
-      double power = json["data"]["16.7.0"];
-      setPowerData(power);
-    }
-    if (json["data"]["1.8.0"].is<JsonVariant>() && json["data"]["2.8.0"].is<JsonVariant>()) {
-      double energyIn = 0.001 * json["data"]["1.8.0"].as<double>();
-      double energyOut = 0.001 * json["data"]["2.8.0"].as<double>();
-      setEnergyData(energyIn, energyOut);
-    }
-  }
-}
-
-double SUNSPEC_scale(int n)
-{
-  double val=1.0;
-  switch (n) {
-    case -3: val=0.001; break;
-    case -2: val=0.01; break;
-    case -1: val=0.1; break;
-    case 0: val=1.0; break;
-    case 1: val=10.0; break;
-    case 2: val=100.0; break;
-    default:
-    val=1.0;
-  }
-  return val;
-}
-
-void parseSUNSPEC() {
-  #define SUNSPEC_BASE 40072
-  #define SUNSPEC_VOLTAGE 40077
-  #define SUNSPEC_VOLTAGE_SCALE 40084
-  #define SUNSPEC_REAL_POWER 40088
-  #define SUNSPEC_REAL_POWER_SCALE 40091
-  #define SUNSPEC_APPARANT_POWER 40093
-  #define SUNSPEC_APPARANT_POWER_SCALE 40096
-  #define SUNSPEC_CURRENT 40072
-  #define SUNSPEC_CURRENT_SCALE 40075
-  #define SUNSPEC_POWER_FACTOR 40103
-  #define SUNSPEC_POWER_FACTOR_SCALE 40106
-  #define SUNSPEC_FREQUENCY 40085
-  #define SUNSPEC_FREQUENCY_SCALE 40086
-  
-  modbus_ip.fromString(mqtt_server);
-  if (!modbus1.isConnected(modbus_ip)) {
-    modbus1.connect(modbus_ip, String(mqtt_port).toInt());
-  } else {
-    uint16_t transaction = modbus1.readHreg(modbus_ip, SUNSPEC_BASE, (uint16_t*) &modbus_result[0], 64, nullptr, String(modbus_dev).toInt());
-    delay(10);
-    modbus1.task();
-    int t = 0;
-    while (modbus1.isTransaction(transaction)) {
-      modbus1.task();
-      delay(10);
-      t++;
-      if (t > 50) {
-        DEBUG_SERIAL.println("Timeout SUNSPEC");
-        //prolong=10;
-        modbus1.disconnect(modbus_ip);
-        break;
-      }
-    }
-    int32_t power = 0;
-    if (t<=50) {
-      double scale_V=SUNSPEC_scale(modbus_result[SUNSPEC_VOLTAGE_SCALE-SUNSPEC_BASE]);
-      double scale_real_power=SUNSPEC_scale(modbus_result[SUNSPEC_REAL_POWER_SCALE-SUNSPEC_BASE]);
-      double scale_apparant_power=SUNSPEC_scale(modbus_result[SUNSPEC_APPARANT_POWER_SCALE-SUNSPEC_BASE]);
-      double scale_current=SUNSPEC_scale(modbus_result[SUNSPEC_CURRENT_SCALE-SUNSPEC_BASE]);
-      double scale_powerfactor=SUNSPEC_scale(modbus_result[SUNSPEC_POWER_FACTOR_SCALE-SUNSPEC_BASE]);
-      double scale_frequency=SUNSPEC_scale(modbus_result[SUNSPEC_FREQUENCY_SCALE-SUNSPEC_BASE]);
-
-      for (int n=0;n<3;n++) {
-        PhasePower[n].power=modbus_result[SUNSPEC_REAL_POWER-SUNSPEC_BASE+n]*scale_real_power;
-        PhasePower[n].apparentPower=modbus_result[SUNSPEC_APPARANT_POWER-SUNSPEC_BASE+n]*scale_apparant_power;
-        PhasePower[n].current= modbus_result[SUNSPEC_CURRENT-SUNSPEC_BASE+n]*scale_current;
-        PhasePower[n].powerFactor=modbus_result[SUNSPEC_POWER_FACTOR-SUNSPEC_BASE+n]*scale_powerfactor;
-        PhasePower[n].voltage=modbus_result[SUNSPEC_VOLTAGE-SUNSPEC_BASE+n]*scale_V;
-        PhasePower[n].frequency=modbus_result[SUNSPEC_FREQUENCY-SUNSPEC_BASE]*scale_frequency;
-        power+= PhasePower[n].power;
-      }
-
-      #define SUNSPEC_REAL_ENERGY_EXPORTED 40109
-      #define SUNSPEC_REAL_IMPORTED_EXPORTED 40117
-      #define SUNSPEC_REAL_ENERGY_SCALE 40123
-      double scale_real_energy=SUNSPEC_scale(modbus_result[SUNSPEC_REAL_ENERGY_SCALE-SUNSPEC_BASE]);
-        for (int n=0;n<3;n++) {
-          uint32_t p=0;
-          uint8_t *p_u8=(uint8_t *)&modbus_result[SUNSPEC_REAL_IMPORTED_EXPORTED-SUNSPEC_BASE+2*n];
-          p|=((uint32_t)p_u8[2])<<0;
-        p|=((uint32_t)p_u8[3])<<8;
-        p|=((uint32_t)p_u8[0])<<16;
-        p|=((uint32_t)p_u8[1])<<24;
-          PhaseEnergy[n].consumption=p/1000.0*scale_real_energy;
-          p=0;
-          p_u8=(uint8_t *)&modbus_result[SUNSPEC_REAL_ENERGY_EXPORTED-SUNSPEC_BASE+2*n];
-          p|=((uint32_t)p_u8[2])<<0;
-        p|=((uint32_t)p_u8[3])<<8;
-        p|=((uint32_t)p_u8[0])<<16;
-        p|=((uint32_t)p_u8[1])<<24;
-          PhaseEnergy[n].gridfeedin = -p/1000.0*scale_real_energy;
-        }
-    }
-    DEBUG_SERIAL.printf("SUNSPEC power: %d,%d\n\r", t, power);
-  }
-}
-
-void queryHTTP() {
-  JsonDocument json;
-  DEBUG_SERIAL.println("Querying HTTP source");
-  http.begin(wifi_client, mqtt_server);
-  http.GET();
-  deserializeJson(json, http.getStream());
-  if (strcmp(power_path, "") == 0) {
-    DEBUG_SERIAL.println("HTTP query: no JSONPath for power data provided");
-  } else {
-    setJsonPathPower(json);
-  }
-  http.end();
 }
 
 void setup(void) {
