+++ src/parsers/MqttParser.cpp
@@ -1,0 +1,21 @@
+#include "../config/Configuration.h"
+#include "../data/DataProcessing.h"
+
+void mqtt_callback(char *topic, byte *payload, unsigned int length) {
+  JsonDocument json;
+  deserializeJson(json, payload, length);
+  setJsonPathPower(json);
+}
+
+void mqtt_reconnect() {
+  DEBUG_SERIAL.print("Attempting MQTT connection...");
+  if (mqtt_client.connect(shelly_name, String(mqtt_user).c_str(), String(mqtt_passwd).c_str())) {
+    DEBUG_SERIAL.println("connected");
+    mqtt_client.subscribe(mqtt_topic);
+  } else {
+    DEBUG_SERIAL.print("failed, rc=");
+    DEBUG_SERIAL.print(mqtt_client.state());
+    DEBUG_SERIAL.println(" try again in 5 seconds");
+    delay(5000);
+  }
+}
