+++ src/parsers/ShrdzmParser.cpp
@@ -1,0 +1,22 @@
+#include "../config/Configuration.h"
+#include "../data/DataProcessing.h"
+
+void parseSHRDZM() {
+  JsonDocument json;
+  uint8_t buffer[1024];
+  int packetSize = Udp.parsePacket();
+  if (packetSize) {
+    int rSize = Udp.read(buffer, 1024);
+    buffer[rSize] = 0;
+    deserializeJson(json, buffer);
+    if (json["data"]["16.7.0"].is<JsonVariant>()) {
+      double power = json["data"]["16.7.0"];
+      setPowerData(power);
+    }
+    if (json["data"]["1.8.0"].is<JsonVariant>() && json["data"]["2.8.0"].is<JsonVariant>()) {
+      double energyIn = 0.001 * json["data"]["1.8.0"].as<double>();
+      double energyOut = 0.001 * json["data"]["2.8.0"].as<double>();
+      setEnergyData(energyIn, energyOut);
+    }
+  }
+}
